Set the initial state of playButton and addButton to false
Set the state of playButton and addButton to true when recording is stopped.

Updated the @IBActoin for the addButton to set 1) the context for CoreData 2) created the sound variable of type Sound 3) set the name and audio attributes of sound.

Saved the context for CoreData to store the recorded audio.

Returned back to the previous View Controller.