Updated the declaration of the ViewController class to include the UITableViewDataSource and UITableViewDelegate.
Created an @IBOutlet for tableView.

Created the property variable named sounds of type Sounds (CoreData class).

Set the dataSource of tableView to self.

Set the delegate of tableView to self.

Removed the didReceiveMemoryWarning function.

Added the numberOfRowsInSection and cellForRowAt indexPath to get rid of error caused after adding UITableViewDataSource and UITableViewDelegate to declaration of the View Controller.

Returned the count of sounds.

Set cell to equal to UITableViewCell ()

Set index of sound to the value IndexPath.row.

Set cell textLabel text to equal to name of the sound.

Returned cell.