Imported the AVFoundation.
Added a property variable name audioPlayer of type AVAudioPlayer.

Added the didSelectRowAt indexPath function.

Added a local constant named sound and set it to equal to equal to the indexPath.row for the property of the variable sounds.

Added a do block to catch and handle the error when trying to play the saved sounds .

Set the value for audioPlayer to equal to AVAudioPlayer (data: sound.audio as! Data).

played the selected audio file using the method play () for the audioPlayer.